Add LINGER support and its testcases.
From FLEXlm manual section 5.2.7 LINGER
LINGER featurename interval
Rarely used. This causes the daemon to `hold on' to the license for featurename for interval seconds after the application checks the license in or exits. This could be useful for short-duration programs which will be used many times in a row by the same user, to ensure that the user will be able to re-acquire the license repeatedly. On the other hand, other users have to wait until the first user is completely finished, plus a linger interval. This is only useful if the application uses duplicate grouping. Otherwise, LINGER will cause you to use extra licenses. Contact your software vendor for information about how they implemented duplicate grouping in their product.

This should finally fix #8.